lmb-892: move the isVisible logic from the application js to the navigationbar js
