+++ app/components/shared/navigation-bar.hbs
@@ -1,4 +1,4 @@
-{{#if @isVisible}}
+{{#if this.isVisible}}
   <AuToolbar @border="bottom" @size="medium" @skin="tint">
     <nav aria-label="breadcrumbs">
       <ul class="au-c-list-horizontal au-c-list-horizontal--small">
