+++ app/templates/application.hbs
@@ -88,7 +88,7 @@
       </li>
     {{/if}}
   </AuMainHeader>
-  <Shared::NavigationBar @isVisible={{this.isNavigationVisible}} />
+  <Shared::NavigationBar />
   <AuMainContainer as |main|>
     <main.content>
       <AuBodyContainer>
